requirements.txt
jinja2
python-frontmatter
python-dotenv
langchain 
langchain-openai
openai
langsmith
instructor
pydantic
fastapi 
uvicorn
requests
websockets